Scrape for new drivers during BSP generation
Previously when a custom driver was added to imx-iotcore it was not
being included in the BuildBSP output, causing feature manifest
failures when the BSP is imported into the iot-adk-addonkit.

This change creates the new folder BSP\{board}\Packages\Misc to store
any additional drivers added to the solution.

1) Scans the driver directory for all .wm.xml files and copies
unrecongnized ones into the BSP.
2) Scans the build output folder and copies all build output from
unrecognized directories into the BSP.
3) Collapses the Packages\Misc folder into a flat structure so the
binaries and .wm.xml files are in the same directory for the package
builder.